refact(server): moved read-only token-standard handlers out of governance.rs
governance.rs mixed the propose -> confirm -> execute state machine with
27 unrelated read-only token-standard/registry/proxy endpoints (47% of
the file), making it the repo's main merge-conflict generator and
defeating review scoping. Moved GovernanceQuery, ContractQueryParams,
and every Services/Configuration/Proxy-tagged handler (vaults, provider
and user services, credential offers/credentials, registrar service
requests/configurations/services, transfer instructions/mint/burn
requests, transfer preapprovals, instruments, transfer factories (plus
its private fetch_amulet_rules_factory helper), holdings, generic
contract query, packages, network-info, operator-info, and
token-standard-contracts) into a new token_standard module, verbatim.

governance.rs keeps the propose/confirm/execute flow, its lifecycle
edges (expire/cancel/cancel-proposal), the audit endpoints, the
coupon-reassignment-delegation read, and the shared helpers
(get_party_token, get_party_credentials, packages) the flow and reward
automation both use.
